bot/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.0
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
psycopg2-binary==2.9.9
table2ascii==1.1.2
wcwidth==0.2.8
yarl==1.9.2
